% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignCluster.R
\name{AssignCluster}
\alias{AssignCluster}
\title{AssignClusterCore needs a shell function around it for formatting, etc.}
\usage{
AssignCluster(lTree, dtDataset)
}
\arguments{
\item{lTree}{is an output of RecursivePartitioningTreeClustering}

\item{dtDataset}{a dataset to assign clusters to. Must have same or more features as lTree}
}
\value{
Returns the rest of the dataset along with node labels at each layer for each row
   NodeLabel Depth Layer1   Layer2     Layer3       Layer4         Layer5           Layer6
1: Root>2>1>2>2>1>1     7 Root>2 Root>2>1 Root>2>1>2 Root>2>1>2>2 Root>2>1>2>2>1 Root>2>1>2>2>1>1
2: Root>2>1>1>1>1>2     7 Root>2 Root>2>1 Root>2>1>1 Root>2>1>1>1 Root>2>1>1>1>1 Root>2>1>1>1>1>2
3: Root>1>2>2>2>1>2     7 Root>1 Root>1>2 Root>1>2>2 Root>1>2>2>2 Root>1>2>2>2>1 Root>1>2>2>2>1>2
4: Root>1>2>2>2>1>2     7 Root>1 Root>1>2 Root>1>2>2 Root>1>2>2>2 Root>1>2>2>2>1 Root>1>2>2>2>1>2
5: Root>1>2>2>2>2>1     7 Root>1 Root>1>2 Root>1>2>2 Root>1>2>2>2 Root>1>2>2>2>2 Root>1>2>2>2>2>1
6: Root>1>2>2>2>1>2     7 Root>1 Root>1>2 Root>1>2>2 Root>1>2>2>2 Root>1>2>2>2>1 Root>1>2>2>2>1>2
}
\description{
AssignClusterCore needs a shell function around it for formatting, etc.
}
