% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignClusterCore.R
\name{AssignClusterCore}
\alias{AssignClusterCore}
\title{Backend for AssignCluster}
\usage{
AssignClusterCore(lTree, dtDataset, iLayerNbr)
}
\arguments{
\item{lTree}{is an output of RecursivePartitioningTreeClustering}

\item{dtDataset}{a dataset to assign clusters to. Must have same or more 
features as lTree}

\item{iLayerNbr}{Keeps track of where in the tree the function is. 
Should initiliase as 1}
}
\description{
Recursive function that traverses the tree and assigns rows of a
dataset to the respective nodes. AssignCluster is the wrapper 
function for this
}
