% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePerformanceFromConfusionMatrix.R
\name{CalculatePerformanceFromConfusionMatrix}
\alias{CalculatePerformanceFromConfusionMatrix}
\title{Calculate performance metric from confusion matrix
Right now, we're just calculating average class error}
\usage{
CalculatePerformanceFromConfusionMatrix(mConfusion,
  cMetric = "AvgClassErr")
}
\arguments{
\item{cMetric}{}
}
\description{
Calculate performance metric from confusion matrix
Right now, we're just calculating average class error
}
