% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{dtUCIVotingRecords}
\alias{dtUCIVotingRecords}
\title{UCI voting dataset}
\format{A data frame with various issues and votes against them}
\source{
\url{www.uci.com}
}
\usage{
dtUCIVotingRecords
}
\description{
Dataset with how people voted on issues. todo.
}
\keyword{datasets}
