% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateNodeCandidateUsingRF.R
\name{EvaluateNodeCandidateUsingRF}
\alias{EvaluateNodeCandidateUsingRF}
\title{Evaluate a feature using random forest}
\usage{
EvaluateNodeCandidateUsingRF(matLabeledDataForClassifier,
  vPredictorFeaturesForClassifier, cPredictedFeature)
}
\arguments{
\item{matLabeledDataForClassifier}{todo}

\item{vPredictorFeaturesForClassifier}{todo}

\item{cPredictedFeature}{todo}
}
\value{
todo
}
\description{
Classifier performance is measured as the average class error on OOB sample
Since RF uses bagging to pick a sample to build each tree, each data point
is unused in roughly 35% of the trees. So the OOB performance basically
takes each point and measures the prediction accuracy on the trees it 
wasn't used to build.
}
