% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateNodeCandidateUsingRpart.R
\name{EvaluateNodeCandidateUsingRpart}
\alias{EvaluateNodeCandidateUsingRpart}
\title{Evaluate a feature using classification tree models}
\usage{
EvaluateNodeCandidateUsingRpart(matLabeledDataForClassifier,
  vPredictorFeaturesForClassifier, cPredictedFeature)
}
\arguments{
\item{matLabeledDataForClassifier}{todo}

\item{vPredictorFeaturesForClassifier}{todo}

\item{cPredictedFeature}{todo}
}
\value{
todo
}
\description{
Classifier performance is measured as the average class error on OOB sample
This is done using 3-fold cross-validation, which is roughly the equivalent
of the OOB performance measured in RF
}
