% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClassificationAccuracyByClassDepth.R
\name{GetClassificationAccuracyByClassDepth}
\alias{GetClassificationAccuracyByClassDepth}
\title{Classification accuracy at various levels}
\usage{
GetClassificationAccuracyByClassDepth(dtTreeSummary)
}
\arguments{
\item{dtTreeSummary}{Output of TraverseTreeToMeasureClassificationQuality}
}
\value{
A summary of the predictions at a particular cut in the tree

   Depth NodeMajorityClass NumRows NodeMajorityPc
1:     2          democrat     104      0.9519231
2:     2        republican     128      0.8046875
3:     2               All     232      0.8706897
4:     3          democrat     104      0.9519231
5:     3        republican     128      0.8046875
}
\description{
At each layer of the tree, calculates the proportion of rows correctly
assigned to a particular label. This is calculated across the effective tree
if you were to cut all the layers off below the specified layer, i.e. all 
leaf nodes upto that layer, and all non-leaf nodes in that layer.
}
