% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewClusteringNodes.R
\name{NewClusteringNodes}
\alias{NewClusteringNodes}
\title{Function to return details of each node in a tabular format. 
TreeStretchedToBottom is a wrapper to this function}
\usage{
NewClusteringNodes(lTree, iParentNodeIndex = 0, iLayerNbr = 0)
}
\arguments{
\item{lTree}{is an output of RecursivePartitioningTreeClustering}

\item{iParentNodeIndex}{= 0 is used to trace the recursive calls. When user is calling 
the function, he/she needn't declare a value for this.}

\item{iLayerNbr}{= 0 is used to trace the recursive calls. When user is calling 
the function, he/she needn't declare a value for this.}

\item{value}{(The members column is wrong in this example. It will actually be members in that node.)
   ParentNodeIndex    NodeLabel         NodeColumnName NodeColumnValue LeafNode Members LayerNbr
1:               0       Root>1          elsalvadoraid               n    FALSE     232        1
2:          Root>1     Root>1>1     handicappedinfants               n    FALSE     232        2
3:        Root>1>1   Root>1>1>1 aidtonicaraguancontras               n     TRUE     232        3
4:        Root>1>1   Root>1>1>2 aidtonicaraguancontras               y    FALSE     232        3
5:      Root>1>1>2 Root>1>1>2>1              mxmissile               n     TRUE     232        4
6:      Root>1>1>2 Root>1>1>2>2              mxmissile               y    FALSE     232        4}
}
\description{
Function to return details of each node in a tabular format. 
TreeStretchedToBottom is a wrapper to this function
}
