% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeOrdinalFeatureChunks.R
\name{OptimizeOrdinalFeatureChunks}
\alias{OptimizeOrdinalFeatureChunks}
\title{Figure out how to divide ordinal features into chunks}
\usage{
OptimizeOrdinalFeatureChunks(mValConfMat, nMinSplits = 2,
  nMaxSplits = 3)
}
\arguments{
\item{mValConfMat}{todo}

\item{nMinSplits}{todo}

\item{nMaxSplits}{todo}
}
\value{
todo
}
\description{
Numeric features can be coarse-classed into bins (e.g. based on quantile 
limits) so that they resemble ordinal features. However, in the interest 
of not having too many splits on a node, adjacent bins may be combined
in order to arrive at fewer splits. If an ordinal feature has $p$ bins
and the maximum number of splits one wishes to have is $q$,
then the number of distinct combinations to be tried out is
$\sum_{r=1}^{q}{C(p-1,r-1)}$ -- this number can be reduced
if one can use domain knowledge to rule out some possibilities.
The optimal split of an ordinal feature can be determined in the same way
that the best features for splitting a node is determined: classifier performance.
}
