% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityOfClustering.R
\name{QualityOfClustering}
\alias{QualityOfClustering}
\title{Measures quality of clustering at various levels}
\usage{
QualityOfClustering(dtDatasetPredicted, cClassLabelFeat)
}
\arguments{
\item{dtDatasetPredicted, }{a result of AssignCluster}
}
\value{
Class     Node   N  Layer       PctOfNode  NodeLabel Correct     LayerCumPct
1: republican   Root>1   5 Layer1 0.04807692   democrat   FALSE 0.021551724
2:   democrat   Root>2  25 Layer1 0.19531250 republican   FALSE 0.129310345
3: republican   Root>2 103 Layer1 0.80468750 republican    TRUE 0.573275862
4:   democrat   Root>1  99 Layer1 0.95192308   democrat    TRUE 1.000000000
5: republican Root>1>2   1 Layer2 0.02439024   democrat   FALSE 0.004310345
6:   democrat Root>2>2   5 Layer2 0.04629630 republican   FALSE 0.025862069
}
\description{
Measures quality of clustering at various levels
}
