% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursivePartitioningTreeClustering.R
\name{RecursivePartitioningTreeClustering}
\alias{RecursivePartitioningTreeClustering}
\title{End function to evaluate (using RF) predictor candidates on which node is 
to be split}
\usage{
RecursivePartitioningTreeClustering(dtDataset, vRows, vNodeCandidates,
  vPredictorFeatures, vKeyFeatures, dtFeatureChunking, lTree, nTreeDepth,
  nMinRows = 1, nMaxTreeDepth = Inf, cClassifier = "RF",
  bKeepFeatureInMixAfterSplit = FALSE,
  bKeepFeatureInPredAfterSplit = TRUE,
  bUseOptimalChunkPerformance = TRUE)
}
\arguments{
\item{bUseOptimalChunkPerformance}{}
}
\value{
todo
}
\description{
Recursive partitioning
This is the wrapper function that recursively partitions the dataset to form
a tree-based clustering solution.
The boundary conditions are:
  - Less than a prespecified number of points at a node
  - No more node candidates
  - No distinct feature values in any of the node candidates
  - Tree depth has reached a maximum
}
