% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagClustersByDepth.R
\name{TagClustersByDepth}
\alias{TagClustersByDepth}
\title{Tag clusters by depth
Iterate through depth levels
For a given depth level, find all the nodes of the tree that are either
nodes at that depth or leaves higher up the tree. Tag them as clusters}
\usage{
TagClustersByDepth(dtTreeSummary, dtDataset, lTree)
}
\arguments{
\item{dtTreeSummary}{todo}

\item{dtDataset}{todo}

\item{lTree}{todo}
}
\value{
todo
}
\description{
Tag clusters by depth
Iterate through depth levels
For a given depth level, find all the nodes of the tree that are either
nodes at that depth or leaves higher up the tree. Tag them as clusters
}
