% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraverseTreeToSummarizeClustersAsDataTable.R
\name{TraverseTreeToSummarizeClustersAsDataTable}
\alias{TraverseTreeToSummarizeClustersAsDataTable}
\title{Summarise tree as data.table with a row for each node
This function will traverse the tree to various depths and create a dataset
with a row for each node in the tree. This dataset will contain details of
how many points that node contains, whether or not it is a leaf node
and the rule that generates that node
Depth-first traversal of the tree happens}
\usage{
TraverseTreeToSummarizeClustersAsDataTable(dtTreeSummary, dtDataset, lTree)
}
\arguments{
\item{dtTreeSummary}{todo}

\item{dtDataset}{todo}

\item{lTree}{todo}
}
\value{
todo
}
\description{
Summarise tree as data.table with a row for each node
This function will traverse the tree to various depths and create a dataset
with a row for each node in the tree. This dataset will contain details of
how many points that node contains, whether or not it is a leaf node
and the rule that generates that node
Depth-first traversal of the tree happens
}
