% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStretchedToBottom.R
\name{TreeStretchedToBottom}
\alias{TreeStretchedToBottom}
\title{Basically adds a Stretch column to the output of NewClusteringNodes}
\usage{
TreeStretchedToBottom(lTree, iParentNodeIndex = 0, iLayerNbr = 0,
  bStretch = T)
}
\arguments{
\item{lTree}{is an output of RecursivePartitioningTreeClustering}

\item{iParentNodeIndex}{Keeps track of where in the tree the function is.}

\item{iLayerNbr}{Keeps track of where in the tree the function is. 
Should initiliase as 1}

\item{bStretch}{Setting it to T ensures that leaf nodes are mentioned in all
layers below the layer of the leaf node as well.}
}
\value{
NodeLabel ParentNodeIndex           NodeColumnName NodeColumnValue LeafNode Members LayerNbr Stretch
    Root>1               0            elsalvadoraid               n    FALSE     104        1   FALSE
  Root>1>1          Root>1 religiousgroupsinschools               n    FALSE      74        2   FALSE
Root>1>1>1        Root>1>1   aidtonicaraguancontras               n     TRUE       1        3   FALSE
Root>1>1>1        Root>1>1   aidtonicaraguancontras               n     TRUE       1        4    TRUE
}
\description{
Basically adds a Stretch column to the output of NewClusteringNodes
}
