% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStretchedToBottomWithDetails.R
\name{TreeStretchedToBottomWithDetails}
\alias{TreeStretchedToBottomWithDetails}
\title{Builds on TreeStretchedToBottom and adds additional details}
\usage{
TreeStretchedToBottomWithDetails(lTree = lTree, dtDataset = dtDataset,
  cClassLabelFeat)
}
\arguments{
\item{lTree}{Output of RecursivePartitioningTreeClustering}

\item{dtDataset}{The original dataset fed to the clustering algo}

\item{cClassLabelFeat}{The name of the column with the assignedclass in the 
original dataset}
}
\value{
NodeLabel ParentNodeIndex           NodeColumnName NodeColumnValue LeafNode Members LayerNbr Stretch NodePrediction Proportion
    Root>1               0            elsalvadoraid               n    FALSE     104        1   FALSE       democrat  1.0000000
  Root>1>1          Root>1 religiousgroupsinschools               n    FALSE      74        2   FALSE       democrat  0.8706897
Root>1>1>1        Root>1>1   aidtonicaraguancontras               n     TRUE       1        3   FALSE       democrat  0.8663793
Root>1>1>1        Root>1>1   aidtonicaraguancontras               n     TRUE       1        4    TRUE       democrat  0.8663793
}
\description{
Builds on TreeStretchedToBottom and adds additional details
}
